Use the node factory rather than operator new
The issue is that when the glTF parser creates QNode subclass
instances using operator new, we always only get the plain old
C++ objects. This is an issue when using QML to drive the
application and when you want to access properties on types that
have been registered with QML as extended types.

The QAbstractNodeFactory will automatically have inner factories
registered at application start up by the QML plugins. That means
we should use the factory to create the nodes rather than operator
new.

This commit does it only for Qt3DCore::QEntity to prove that it
works.

Without this commit accessing a Kuesa.Asset.node in QML code only
gives access to the basic QNode properties. With this commit we
get access to the properties defined on the QEntity extension
object too - namely the list of components. We can then use this
to mutate the entity by adding more components such as an
object picker to integrate the scene with our application logic.