Perform Exposure, Tone Mapping and Gamma Correction as last step of the pipeline
This allows FX to work in linear RGB space and leave out any of the above
conversions out. This gives better control over the order of the FX and ensure
we introduce lossy conversion as late as possible.

Also use RGBA16_UNorm color attachments so as not to lose HDR precision.

Only downside is we now need a FBO where we didn't need one when not using FX
previously.

Materials graphs edited accordingly.

Task-Id: KUE-756
Change-Id: I5a117781304b75cebce13e278bc2d5916e40ca59
Reviewed-on: https://kuesa-codereview.kdab.com/c/kuesa/kuesa/+/169
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>