Doc: fixes to work offline
The bootstrap css file was being pulled from an online source.
Although this is normally okay, it could cause issues down the
line for our scenario. This changes Kuesa docs to include the
file and now refer to it.

Also add assets (style/images) to qhp.extraFiles for proper
QtCreator doc generation

Change-Id: I797112dc7c2cfba4fd20c4f1dc30d50b981cff12
Reviewed-on: https://kuesa-codereview.kdab.com/c/kuesa/kuesa/+/307
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>