Custom Materials support
The idea is to make it as easy as possible to add new custom materials.

This will require having glTF2 exporters use the KDAB_custom_material
extension which will specify the value of the custom material's
properties and the type of material.

On the Kuesa side, this will mean we have a matching implementation of
the material with the matching properties. To ease that process, a
python class generator reads a JSON based description of the material
and generate the relevant classes.

Custom material classes can be registered against the GLTF2Importer
using GLTF2Importer::registerCustomMaterial. This makes it easy to
add new custom materials as part of Kuesa but also as part of a
specific application.

Task-Id: KUE-825
Change-Id: I1ee348035418f47c7796cb4824509eb8db8c1baa
Reviewed-on: https://kuesa-codereview.kdab.com/c/kuesa/kuesa/+/345
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>