Improve depth of field.
Fix some issues with the depth of field where blurred
object in background clobbered in-focus objects. Also fix
problem where everything was blurred to some extent except
for points exactly along the focal plane.

To get a good depth of field we need 2 textures:
1. the original unblurred sharp scene representing the in-focus
   area of the scene.
2. A blurred texture where things are blurred based on the
   circle of confusion, determined by distance from focal plane.

When calculating the blurred texture, we also need to determine
whether a point is in the foreground or background.  This is
because objects in fron always obscure objects behind them.  Thus
sharp objects are not blurred by out of focus object behind them.
Blurred objects in the foreground do blend into in-focus objects and
blurred objects in the background. We store the background vs foreground
amount in the alpha channel of the blurred texture.

DoF is now done use 2 render passes.  The first to create the blurred
texture and fg/bg ratio. The sharp texture is the original unblurred scene.
The 2nd pass combines these, using the sharp texture everywhere the
circle of confusion (CoC) is smaller than a certain threhold. Otherwise
it uses a ratio of the blurred texture and sharp texture determined by
the foregroundness of the point in the blurred texture. The more
background it is, the less it'll blur into the sharp texture. If it's
more foreground, it will blur into the sharp texture so it samples
mostly from the blurred texture.

Possible improvment:
Add 3rd pass to calculate the CoC and store in texture. It would clean
up duplicated CoC calculation. Not sure whether it would make things
faster or not.  would still need to do same number of texture reads
when blurring but not all the calculations associated with CoC.

Task-Id: KUE-851
Change-Id: Iac12a96f2f06f11308960009ee4a64d7ec8265e8
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/75030
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>