Fix ugly deformation of slider control handles in UI
The problem is that the border images used don't scale down well. They
have a size of 33 pixels and are meant to not scale to under that size.

I have replaced the use of the images for the handle with using a rounded
rectangle with a border instead. This gets rid of the issue. The colors
used for the rectangle have been picked from the original image.

The dot has also been re-centered a bit, to take into account its position
needs to change a bit if the height of the containing handle is smaller than
the intended 33 pixels. Otherwise, the dot appears more to the left of the
handle if the slider has a value of 0.

Change-Id: I7b3ee7833f8e458fc9d115dd716014153812812e
Reviewed-on: https://kuesa-codereview.kdab.com/c/kuesa/kuesa/+/382
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>