Fix multi pass transparent drawing
We need to draw all the passes of a multi pass transparent material one after
another to ensure we maintain a proper back to front rendering order.

Therefore using multiple RenderPassFilter is not option. We can instead just
use a single RenderPassFilter. Adding multiple RenderPasses with the same
filter key to a material results in the entity being drawn with all the passes
that match the filter in the order in which they were added.

Change-Id: I48167d572ea106711d789f35ef9748b35fb5ac84
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/77991
Reviewed-by: Mike Krus <mike.krus@kdab.com>