Update shader graphs and the iro-materials-gallery example
materials: A new gltfYUp property was added to materials which care about which
axis is up. If set to true, it means the scene was designed in a Z-up
environment and transformed to Y-up on export.

iro-materials-gallery: All kuesa nodes were re-instanciated.

gltf files for custom Material unit tests were also updated

Change-Id: Icd5b61facafe0ab95df4b342046d8e6f08239121
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/78132
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>