Work around bug in Qt 3D 5.12 for non-transform animations
Qt < 5.14 has a bug where change messages will not be delivered to
frontend nodes if these don't have a backend equivalent. This prevents
animation changes from being delivered to the frontend.

Work around this by creating a dummy observer and registering it with
the change arbiter.

5.14 and up don't have this problem as messages are no longer used
and all changes are propagate via the backend.

Change-Id: I1a87c707247eeaae7ee0e29e6a446a2e1b68a442
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/78241
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>