IroGlassAddMaterial: add diffuseInner/OuterProperties
Since this material overloads IroDiffuse, we need to expose those properties to
be sure that we initialize the uniforms.

Failure to do so could result in the uniforms remaining uninitialized which
could result in a potential undefined behavior depending on how OpenGL handles
this. Even if OpenGL assumes a 0 default value, we could end up with the uniform
getting the value set by a previous draw calls if using the same shader and
previous draw call had uniform value different than 0 (which can happen since
IroDiffuse and IroGlassAdd use the same shader).

Change-Id: I2d5703890dadfe1692d740e3bf8249efde0e95ad
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/79250
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>