Implement shadows in Kuesa
This patch has working hard and soft shadows for all 3 light
types supported in Kuesa. All opaque materials cast and
receive shadows. This works currently on Desktop GL and
needs shaders for es 2 and es 3

Create a ShadowCastingLight class and make the Kuesa lights
inherit from it. This is a good place to indicate whether a,
light casts a shadow and to put all of the required parameters
and shader uniforms except for the shadowmap textures.

The shadowMap textures are part of the GLTF2 Material instead
of the lights because otherwise we'd need 2 textures samplers
per light times 8 lights which uses too many samplers. We
rely on Qt3D to popuplate the Light uniforms for us so we
can't change this. By moving the textures to the material
we can use 2 array textures.

Add a ShadowMap class that takes a ShadowCastingLight and has
the necessary camera info to render a shadowmap based on light
position and type. This class updates the ShadowCastingLight's
uniforms based on shadowmap camera parameters.

Add a ShadowMapManager that listens to any lights being added
to the LightCollection and creates ShadowMaps for enabled lights
that cast shadows. It uses the EntityTransformWatcher class
to obtain the world position and direction for each light.
It also registers a QFrameAction that to get the scene bounding
volume which is used to set the camera parameters (near/far,
position for directional lights, etc) when rendering the
shadowmap.

Update ForwardRenderer to listen to the ShadowMapManager's
list of shadowmaps and add the required FrameGraphNodes to
render all the shadowmaps.  Each light/shadowmap requires a
separate render pass using a camera based on the light position
and direction. Use the z-fill shader since it's simple and all
we need is depth information.  We only render opaque objects
into the shadowmap. Point lights are more complicated as they need
to be rendered into a depth cubemap. This can be done in a single
pass using a geometry shader.

For now the ShadowMapManager lives in the LightCollection
and lights must be added to the collection for shadows to show
up.  Also, there is a manual step of setting the
ShadowMapManager on the ForwardRenderer...seems like this
should happen automatically somehow.

Created a shader file that can be included to add shadows to any
material shader fairly easily.

Added shadows to the manual light test and added 2 lights of each
type.  Also  added a few more gltf test files.

Future Improvments:

* shaders for ES2/3
* indicate whether material casts shadows
* update iro materials
* Improve shadowmap camera parameters. In particular, allow the user
to set either a bounding volume for shadows per light or per scene, or
to provide an entity to use as the bounding entity. We try to calculate
this automatically, but it's not great. For directional lights
in particular we are rendering a shadowmap for entire scene which
isn't good if only viewing a small subset.

[ChangeLog] Implement shadows in Kuesa

Task-Id: KUE-946
Change-Id: I26bf79c6a09422b46c3baf0d971bea7e1bdb9294
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/82031
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>