Fix bug enabling and disabling lights and add test
We need to add all of the shader uniforms to each light's
QShaderData regardless of whether that light type needs it. Qt3D
automatically adds all of its light's uniforms, but not the extras
we added for Kuesa.  If we start a scene with one light type
enabled, Qt3D will match that light's uniforms to the shader but leave
out any shader uniforms that it doesn't use. Enabling a different light
type will fail if that light has uniforms not set by the 1st light.

For now just add them in the constructors. We'll likely add a new
base class when we add shadows and move the initialization there.

Added a simple manual test to toggle between all 3 light types.
Will likely improve this test as part of shadow work.

Change-Id: I6d44fef34bef9d29c2accafae2d9cb008940daac
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/84639
Reviewed-by: James Albamont <jim.albamont@kdab.com>