MSAAResolver: set parent after instance is created
This fixes an assert/crash that occurs with Qt6. It seems that the LayerFilter
instance created in the MSAAResolver ctor doesn't get properly registered with
the Qt3D backend (likely due to the deferred nested contruction). This results
in an assert being triggered when the MSAAResolver node gets destroyed which
occurs when the ForwardRenderer framegraph is reconfigured.

This can be easily reproduced by enabling and disabling the OpacityMask effect
on the car-scene example.

Change-Id: I91237d14c657c9654a76d18a3c1fd55a1acb3d7f
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/84810
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Mike Krus <mike.krus@kdab.com>