Add basic planar reflection support to ForwardRenderer
The rendering goes as follow:
1) Render scene reflected against a plane normal (straight to color buffer)
2) Clear Depth Buffer
3) Render scene from camera point of View

This required a slight refactoring of the way we handle rendering stages.
We now have the ReflectionStages which perform the reflected rendering
of the non frustum culled scene followed by the clearing of the depth
buffer. Next, we perform the SceneStages to render the scene from
the current point of view.

The reflection content is not clipped and no fall off is applied.
This will be improved upon in following patches.

Note: RHI doesn't allow to only clear the depth without also clearing
the color buffers. This might force us to render this step into a
texture and composite it later on in the scene.

Task-Id: KUE-1074
Change-Id: Ie01c8ef1b98362248fe26cb971bbfcf8a1c35dee
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/85189
Reviewed-by: Mike Krus <mike.krus@kdab.com>