Add FrameGraph View
A view is a FrameGraph representing a view of the scene.
It can either be the full scene or a subset of the scene.

The view allows to specify which rendering features are required.

The ForwardRenderer will be updated to become a subclass of View
and contain will contain an optional list of views. This will allow
to handle both cases of having a single view of the scene or multiple
ones.

Additionally, reusable pieces of the FrameGraph tree will be moved
into standalone Stages classes. This will make it a lot simpler to
maintain/understand and will give us a way of easily testing individual
parts of the FrameGraph at the expense of some FrameGraph nodes duplication.

Tests, documentation and class are completed in the following commit.
Keeping this commit separate to make reviewing process slightly easier.

Change-Id: I624e3bd557a85cfcf15b70d11de07732b47a3ad1
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/85232
Reviewed-by: Mike Krus <mike.krus@kdab.com>