Register all QML types against version 1.0
We haven't observed any source compatibility with Kuesa.
This apparently seems to lead to strange bugs when registering various revisions
of a same class, especially if it starts subclassing other classes that didn't
exist or weren't registered in earlier versions.

For the above reason, we now register everything against 1.0, which will allow
any import with subsequent minor versions (as long as we have an appropriate
qmlRegisterModule)

Add qml tests to check type registration and instantation and fix duplicate
property on UnlitEffect.

Change-Id: Ie4d70dece0cca853df73ce9b8236ccba7f7933b2
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/86474
Reviewed-by: Mike Krus <mike.krus@kdab.com>