Fix Bloom FX being flipped with the Vulkan Backend
The Bloom FX is actually composed of 3 internal FX

- A single pass threshold effect
- A n * 2 passes gaussian blur effect (when n >= 1)
- A final bloom pass that addes an input texture to the blurred and thresholded
  texture created by the previous passes.

What happens is that since the thresholding and gaussian blur passes leads to an
odd number of passes, when we get to the final bloom pass, the generated
texture is always flipped the other way than the input texture.

To work around this, we now expose a yFlip setter on the Threshold Effect to work
around that.

Change-Id: If504f3dfe4210ebe077b744fa45c1b7b0a4dfdb3
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/86839
Reviewed-by: Mike Krus <mike.krus@kdab.com>