Add AssetCache
Will be used to keep a cache of Qt3D resources for the importer.
If the GLTF resource it is associated with provides a key attribute,
the resource could then be cached which would allow to reuse some
assets when changing the source of the importer.

The AssetCache provides will take care of parenting the resource to
a sceneEntity to avoid them being destroyed when changing scene.
It also watches for Resource destruction so that they get removed
from the cache when that occurs.

Change-Id: I293fb4ecaddebe787f976b43c538c9f78245da3e
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/86904
Reviewed-by: Mike Krus <mike.krus@kdab.com>