Fix destruction order of effects
The FXStage should unparent the fx fg subtrees upon destruction so
that effects are deleted when the sharedptr that last references them is
destroyed and not when their parent FG node is destroyed.

ForwardRenderer was outdated and referencing an empty set of FG tree
since the recent View and EffectsStages introduction.

This fixes crashes on exit of the car-scene example when fx where enabled.

Change-Id: I65bb5f033f8f15a9c5ba96b1d95a2644df89a3e2
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/87352
Reviewed-by: Mike Krus <mike.krus@kdab.com>