ForwardRenderer::removeView fix accessing destroyed object
removeView can be called as a result of a View instance being destroyed.
In that case, the View parameter we receive is a partially destroyed
object and we should not try to access it. Yet in the case the view
is removed but not destroyed, we still need to ensure that the FG
stages are properly un parented. This is now done directly in
reconfigureFrameGraph() by iterating over all children of the
scene stage and effect stage root nodes.

This fixes crashes when running tst_ForwardRenderer

Change-Id: I1ee71a20f1749ade646d36ac6b91c731c03b5cd8
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/88353
Reviewed-by: Nicolas Guichard <nicolas.guichard@kdab.com>