Make view refactoring work with RHI
Once again, we get hit by the fact that RHI doesn't allow to
draw without clearing the back buffer.

To work around this, instead of having each View draw directly
to screen, we render the last pass of each view into the
already allocated view render target.
Then, we have a final pass that iterates over each View
and renders the final render target of each View on screen.
This approach works because Qt3D groups all RV that target the
same target together when they are in a sequence.

Change-Id: I71b43916c840033799031127dd322dc3136d19e7
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/89118
Reviewed-by: Mike Krus <mike.krus@kdab.com>