Improve View3DScene API
By making source and cameraName readonly properties.
They can now only be set from the activeScene. This avoids the
confusion of having a case where source and cameraName would
be set from both the properties and the activeScene.

For most cases, this change should be transparent for uses
as View3D hides that change.

Also fix export uri for SceneConfiguration in QML

Change-Id: I5555c640cac8976ebb4cc2ae51aa605aa6ba6492
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/89173
Reviewed-by: Mike Krus <mike.krus@kdab.com>