Take ShadowMapManager private and auto-connect to ForwardRenderer
Make the ShadowMapManager private as it is an internal detail of
Kuesa. The ForwardRenderer needs to know when the ShadowMaps change,
which used to be done by setting the ShadowMapManager as a property
on the ForwardRenderer. This was clunky and required every use of
ForwardRenderer to remember to do this.

Fix by by connecting the ShadowMapManager's signal to the ForwardRenderer.
This is a bit trickier than it seems since we never really get a pointer
to the ForwardRenderer. It's often declared in QML as a property on
the RenderSettings.  Use Qt3D's private API to get the backend FrameGraph
root node and use it to look up the corresonponding QNode in the QScene.
We need to do this once the scene is actually rendering, so use
a single QFrameAction.

If this patch is OK, will squash with actual Shadows implementation
patch.

Change-Id: Ie4d9e722699dc815a37692c0cdf49e270bc4538d
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/89362
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>