Introduce PlaceholderTracker to watch Placeholder entities
Let Placeholder only be a positioned entity in the scene tree. To watch a
Placeholder, use the PlaceholderTracker which automatically retrieves a
Placeholder instance from the sceneEntity collection based on a name. It also
takes care of performing the actual computations based on the Placeholder
transform, a camera, viewport and screen size.

Another benefit of decoupling viewport/camera from the Placeholder is that we
can track the same placeholder from different cameras and viewport.

It works essentially the same as the more generic TransformTracker
and uses the same property names and usage.

View3D and View3DScene updated to have a list of PlaceholderTracker
on which camera and scene size are automatically set.

[ChangeLog] Introduce PlaceholderTracker to watch Placeholder

Change-Id: Icceef8cf998f18054d07c692513e6818b46ee755
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/91514
Reviewed-by: Mike Krus <mike.krus@kdab.com>