Drill-Experience: make AbstractScreenController a QObject
So that the sceneConfiguration can be removed from the Qt 3D
scene tree when we switch to another controller's sceneConfiguration.
Otherwise, the sceneConfiguration even if not the active one
remains parented and therefore any of its assets are still picked up
by Qt 3D and possibly making unwanted changes in the 3D view.

Change-Id: I576464a7ea612de514110e05aace9d0aa011bd73
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/92852
Reviewed-by: Nicolas Guichard <nicolas.guichard@kdab.com>
Reviewed-by: Juan Casafranca <juan.casafranca@kdab.com>