View3DScene: fix parenting of SceneConfiguration which had no initial parent
View3DScene automatically parents a SceneConfiguration which has no parent
or whose parent is not a QNode. Upon setting a new active scene, the
original parent is restored on the previous scene. However the old code
was not setting the parent back to null in the case there was no initial
parent set. This patch corrects that.

This behavior is indeed useful to ensure that assets parented by the
SceneConfiguration don't remain part of the Qt 3D scene tree when switching
active scene configuration.

Change-Id: I1c88b646baeb6a8abee9d9eea3641bdb6c3f7dde
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/92766
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>