ForwardRenderer: fix dynamic addition/removal of Views
It appears that when adding multiple views at once during runtime,
Qt 3D doesn't always pick the changes correctly.

This might be because of the adding/removing/adding... of nodes
sequence when several Views are added. For some reason, having intermediate
FG nodes to hold the Views rather than directly parenting to the
surface selector works.

Note: also removed the < 5.15 code paths which are not needed anymore
and make the code and the tests harder to understand.

Change-Id: I99cf7ee0589106e051b2f7bb7c3ac31898c21061
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/92944
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>