View3DScene: clear and reload on activeScene switch
Even if the source being referenced by the SceneConfigurations is identical.
This is so that we can ensure a clean state of the loaded scene content not
impacted by prior animation or manual property changes.

To give the SceneConfiguration instance a chance to react to the fact it was
unloaded, a new unloadingDone signal was added to it.

Task-Id: KUE-1158
Change-Id: If2803d180273427a568312d93467407dad9b2fe5
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/93305
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>