Remove QRenderPassFilter in TransparentRenderStage
Transparent render stages are defined by two passes and are not
conmutative nor associative. This may be an issue if we have
two transparent objects one in front of another, since we first draw
pass0 of both objects and then pass1 of both objects, which gives
different results to what we expect and have in blender.

Also remove the Material sorting policy as that groups render passes
together, which may rearrange the order in which the render passes
are executed.

Change-Id: I13bf317fd9f79e076c247905c6cf602577536f01
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/94656
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>