View: enable back to front sorting for transparent materials by default
We had it disabled by default as a way to ensure the best performance. However
in most cases, when dealing with transparent objects, we want it to be enabled.
Therefore, it makes more sense to have it enabled by default and let people with
advanced use cases or performance issues disable it if needed.

This fixes flickering of transparent objects with iro-materials-gallery with Qt6
following the change in RenderPass filtering.

[ChangeLog]: enable back to front sorting for transparent materials by default

Change-Id: I3de5f5797901f7f5cb7df71c8ad9837a17e86012
Reviewed-on: https://codereview.kdab.com/c/kuesa/kuesa/+/94804
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Juan Casafranca <juan.casafranca@kdab.com>